+++ jscomp/runtime/js_undefined.ml
@@ -32,12 +32,10 @@ open struct
   module Js = Js_internal
 end
 
-external to_opt : 'a t -> 'a option = "#undefined_to_opt"
 external toOption : 'a t -> 'a option = "#undefined_to_opt"
 external return : 'a -> 'a t = "%identity"
 external empty : 'a t = "#undefined"
 
-let test : 'a t -> bool = fun x -> x = empty
 let testAny : 'a -> bool = fun x -> Obj.magic x = empty
 
 external getUnsafe : 'a t -> 'a = "%identity"
@@ -47,7 +45,8 @@ let getExn f =
   | None -> Js_exn.raiseError "Js.Undefined.getExn"
   | Some x -> x
 
-let bind x f = match to_opt x with None -> empty | Some x -> return (f x [@u])
-let iter x f = match to_opt x with None -> () | Some x -> f x [@u]
+let bind x f =
+  match toOption x with None -> empty | Some x -> return (f x [@u])
+
+let iter x f = match toOption x with None -> () | Some x -> f x [@u]
 let fromOption x = match x with None -> empty | Some x -> return x
-let from_opt = fromOption